env command works
