+++ Makefile
@@ -35,7 +35,7 @@ OBJDIR_TEST = objs_tests
 SRC = main.c fn_msh_start.c fn_msh_parser.c fn_msh_executer.c fn_msh_clean.c
 SRC_LEXER = lexer_list_utils.c lexer_utils.c lexer.c lexer_token_env.c
 SRC_PARSER = parser.c
-SRC_EXECUTER = executer.c redirect.c
+SRC_EXECUTER = executer.c redirect.c execute_env.c execute_export.c
 SRC_ENV = init_env.c expansion.c
 SRC_SYNTAXER = syntaxer.c syntaxer_utils.c
 SRC_TEST = test.c test_lexer.c
