+++ include/executer.h
@@ -6,7 +6,7 @@
 /*   By: jguacide <jguacide@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/04/08 15:22:14 by jguacide      #+#    #+#                 */
-/*   Updated: 2025/04/08 15:22:15 by jguacide      ########   odam.nl         */
+/*   Updated: 2025/04/08 22:18:04 by sveta         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,6 +32,9 @@
 void execute_commands(t_minishell *mshell);
 void handle_exe_redirections(t_command *command);
 void execute_single_command(t_command *command);
+void exec_export(t_command *command, t_minishell *mshell);
+void exec_env(t_minishell *mshell);
+
 int is_builtin_command(char **command_args);
 void execute_builtin(char **command_args);
 void restore_file_descriptors(int saved_stdin, int saved_stdout);
